requirements.txt
ovh~=1.2.0
pandas~=2.2.3
docker~=7.1.0
requests~=2.28.2
setuptools~=75.1.0
python-dotenv~=1.0.1
